+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 
 ### Features
 
-* store, volume: two-phase compaction with segment rewrite, configurable default namespace, WAL replay support ([dc09b6c](https://github.com/asaidimu/blobs/commit/dc09b6cefcfad9a0510d2a1bf73648c3f79287cc))
+* store, volume: two-phase compaction with segment rewrite, WAL replay support ([dc09b6c](https://github.com/asaidimu/blobs/commit/dc09b6cefcfad9a0510d2a1bf73648c3f79287cc))
 
 ## [1.0.2](https://github.com/asaidimu/blobs/compare/v1.0.1...v1.0.2) (2026-07-15)